feat: add ability to have a composite key
This adds the ability to define composite keys for identifying records.
For example, you can define field indices 0 and 1 to be the key,
so we end up with the following:
```
header1,header2,header3,...
a,b,c
d,e,f
```
In this example a,b identifies the first record, while d,e identifies
the second.