perf(hash): fill hasher with as few operations as possible
When creating the hash for a csv record, it has iterated through
every field separately and wrote to the hasher.

Now, it takes the whole record as a slice and writes it to the hasher
*once*.
This leads to an incredible performance improvement of nearly 20%!

Keep in mind that this mechanism will slightly change/get more
complicated as soon as we introduce the possibility of *ignoring fields*
for the comparison.