wip: it now basically works
API is still not 100% perfect, because we had to make some internals
`pub`, but it is now usable.
Names of traits are still bad -> need to refactor

Also we have to check, whether we can move some impls to the more
generic CsvDiff<S, T>, because otherwise all this effort is in vain.