fix: replace `unwrap`s with proper error handling when reading csv
This implements proper error handling when reading csv records.
In doing so, it uses the error type of the `csv` crate and propagates
it up to the caller.

In future iterations, we have to look into wrapping the `csv` error type
into our own error type, because we will probably introduce errors such
as:

- (CsvLeftError, CsvRightError)

meaning we try to parse all errors instead of stopping at the first
error and return that. Furthermore, errors in `csv` also don't tell you
anything about *in which csv* (e.g. left or right csv) the error
occured, so this is something to improve on in the future.