feat(diff_result): add proper type for sorted `DiffResult`
This implements an enum that indicates the diff result is sorted.
However, this api has a huge problem:
We currently only expose an api for sorting by line (so there exists one
enum variant). But when we want to expose further sorting mechanisms
(e.g. sort by `DiffType`), we will have to add another enum variant,
which is a breaking change.
So this api will probably change in the future (or rather: in the next
commit ;-)).