refactor(diff_row): use `Vec` instead of `HashSet` for field indices
The changed field indices are best represented by a `Vec`, because they
have an ordering. With `HashSet` we would loose that ordering.

The implementation/algorithm for filling the collection doesn't need to
change, because each field index that is read is already in the right
sequence and every field index only occurs once, so there are no
duplicates.