feat: introduce `Csv<R>` newtype
This introduces a new main type `Csv<R>`, which wraps the former
`Read` item that was send to the hashing tasks.

The advantage of this is that we can configure the `csv::Reader` through
`Csv<R>`.
At the moment, only the boolean option `headers` is taken into account,
but this will be extended in the future to account for all config
options `csv::Reader` provides.