fix(hash): use xxhash with 128-bit length
This fixes the hashing of csv records by using `xxhash` with 128-bit
digest length.
Before, `ahash` has been used, which only has 64-bit digest length.
According to the birthday paradox, this would lead to collisions with
a probability of 50%, when we hash 2^32 different items, which are too
few items.
With xxhash and 128-bit digest length, we now achieve 50% collision
probability when hashing 2^64 items (provided that the hash evenly
distributes, which seems to be the case:
https://github.com/Cyan4973/xxHash/wiki/Collision-ratio-comparison).
For further background and discussion on this change, please see:
https://users.rust-lang.org/t/using-ahash-non-cryptographic-hash-for-equality-comparison-considered-harmful/71701