feat: implement proper streaming
We can now properly stream csv hashes (in a push-based model) via
an iterator.

This also introduces a new dep called `io_arc`.
It allows us to wrap something that can `Read` into an Arc,
without loosing the ability to `Read`.

There is still _a lot_ to do, before we can stabilize this, but it
is a major step, nevertheless.

For example we need to use `VecDeque` instead of `Vec` in the iterator.