fix(perf): start receiver thread first
This improves the performance of the streaming diff by moving around the
order in which threads start to be spawned:
Before, the receiver thread started last, which led to suboptimal
performance due to the fact that a lot of csv records have been sent
into the channel without being received yet.
Now, the receiver thread starts first and is ready as soon as the sender
threads start sending csv records, so they can be immediately compared
(and sent back upstream to reuse memory, if records are equal).