test: refactor tests => remove duplication
This refactors the tests, so that we have less code duplication in the
tests.

It also adds tests for the streaming mode (`CsvByteDiff`).

The tests are now structured in a way, where `CsvByteDiff` and
`CsvByteDiffLocal` are tested in the same method.

Note, however, that not all tests cover `CsvByteDiff` yet, especially
those that need a builder, because there is no builder for this struct
yet.