wip: calc max dynamic threshold for map
This implements the calculation of the max dynamics threshold for the
hash maps.
Currently, the dynamic variant is always used. This will change in the
future, when we integrate the "approx_num_of_lines" calculation (then,
static variant will be used).

Warning: there is still a huge bug: the parsing that happens in `Csv`
for the first few lines will discard all further lines that need to be
still parsed! We have to investigate on how to solve this.