wip: fix parsing first few csv lines
This fixes the parsing of the first few csv lines. The problem was
that it sometimes parsed beyond the end of the csv, because we didn't
look at the boolean return value of `read_byte_record`.
This resulted in an empty `ByteRecord` and therefore an `Add`ed or
`Delete`ed record in the diff result.

There is still a deadlock happening when comparing >= 10,000 lines.
This requires further investigation.