fix: fix deadlock when diffing >= 10,000 lines
The deadlock happened, because the receiver side of the "first few"
channel blocked waiting for the sender side to be dropped, but it was
never dropped, because the channel for the "remaining" lines wasn't
finished yet, but has reached it's max capacity of 10,000 elements and
so was blocked, too, which prevented the "first few" channel from being
dropped.
We now drop the "first few" channel explicitly after it is done with
sending.