wip: provide size hint
This enables us to provide a size hint on how many csv rows are to be
expected, so that we can optimally reuse memory.

It does this by parsing the first 5 lines of CSV and calculating the avg
byte count per line.
If simultaneously a total byte count of the csv is given, then we can
calculate the expected total number of lines of the csv.

The parsing of the first 5 lines is special here, because it differs
from the normal parsing process: the normal parsing process does all
sorts of trickery to reuse memory (e.g. through the `recycle_receiver`),
whereas this optimization is not done for parsing the first 5 lines
(to reduce complexity).

Warning: there is still a huge bug: the parsing that happens in `Csv`
for the first few lines will discard all further lines that need to be
still parsed! We have to investigate on how to solve this.