fix: handle csv errors in `CsvDiff`
This handles csv errors in `CsvDiff` that can occur, when parsing csv.
Before, errors where just "swallowed"/ignored and so the iterator just
stopped.
Now the iterator returns a `csv::Result`, so that errors can be handled
properly on the call site.

Note, that this is hurting performance a bit:
before: ~710ms for 1,000,000 rows
now: ~780ms for 1,000,000 rows