fix: take `Arc<ThreadPool>` instead of 'static ref
When constructing a `CsvDiff` with a builder, there is the possibility
to pass it a `rayon::ThreadPool`. But it accepted a 'static ref, which
does not live long enough.
Now, it accepts an `Arc<rayon::ThreadPool>`, which resolve the borrowing
issues and should be in alignment in the greater context of using a
thread pool (thread pool will be shared among many components via
`Arc`).