Add `sort_by_columns` method on `DiffByteRecords`
This implements a new method on `DiffByteRecords` that allows to sort
by columns according to the result of comparing two &[u8]'s.

The compare logic is as follows:
- always compare against the byte record of `DiffByteRecord::Modify`'s
`delete` row
   - if the result is equal, compare against `DiffByteRecord::Modify`'s
   `add` row
- When comparing two `DiffByteRecord::Modify` rows with each other
   - compare `DiffByteRecord::Modify`'s `delete` row from the first with
   the `DiffByteRecord::Modify`'s `delete` row from the other
   - if the result is equal, compare instead both
   `DiffByteRecord::Modify`'s `add` rows with each other
- if the actual byte records are equal and every column has been
compared as equal, the following relation holds
   - `DiffByteRecord::Delete` < `DiffByteRecord::Modify` < `DiffByteRecord::Add`

The new method `sort_by_columns` can fail, if indices are provided that
are out of bounds. In that case, the first error will be cached and
return from the method until sorting is complete. If an error occurs,
the method still tries to finish the sort (e.g. there might still be
some indices left that will decide the sort order, because the previous
indices led to columns being compared as equal). But it is recommended
to not rely on the sort order, if an error has occured (sort order is
basically undefined then).